requirements.txt
streamlit
pandas
openpyxl
requests
gspread
openai
oauth2client
python-dotenv
watchdog
